requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bitarray==2.9.2
cachetools==5.4.0
cbor2==5.6.4
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
crytic-compile==0.3.7
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
fastapi==0.112.0
frozenlist==1.4.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.2
grpcio-status==1.62.2
h11==0.14.0
hexbytes==0.3.1
httplib2==0.22.0
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
packaging==24.1
parsimonious==0.10.0
prettytable==3.10.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.19.1
rsa==4.9
slither-analyzer==0.10.3
sniffio==1.3.1
solc-select==1.0.4
SQLAlchemy==2.0.31
starlette==0.37.2
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
wcwidth==0.2.13
web3==6.20.1
websockets==12.0
yarl==1.9.4
